+++ core-ng/src/main/java/core/framework/api/web/exception/ConflictException.java
@@ -1,12 +1,13 @@
 package core.framework.api.web.exception;
 
+import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Warning;
 
 /**
  * @author neo
  */
 @Warning
-public class ConflictException extends RuntimeException implements ErrorCode {
+public final class ConflictException extends RuntimeException implements ErrorCode {
     public static final String DEFAULT_ERROR_CODE = "CONFLICT";
 
     private static final long serialVersionUID = 7787085179989898162L;