+++ core-ng/src/main/java/core/framework/api/web/exception/MethodNotAllowedException.java
@@ -1,13 +1,14 @@
 package core.framework.api.web.exception;
 
 import core.framework.api.http.HTTPMethod;
+import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Warning;
 
 /**
  * @author neo
  */
 @Warning
-public class MethodNotAllowedException extends RuntimeException implements ErrorCode {
+public final class MethodNotAllowedException extends RuntimeException implements ErrorCode {
     private static final long serialVersionUID = 2349080664326196294L;
 
     public MethodNotAllowedException(HTTPMethod method) {