+++ core-ng/src/main/java/core/framework/api/web/exception/NotFoundException.java
@@ -1,12 +1,13 @@
 package core.framework.api.web.exception;
 
+import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Warning;
 
 /**
  * @author neo
  */
 @Warning
-public class NotFoundException extends RuntimeException implements ErrorCode {
+public final class NotFoundException extends RuntimeException implements ErrorCode {
     private static final String DEFAULT_ERROR_CODE = "NOT_FOUND";
     private static final long serialVersionUID = 8663360723004690205L;