+++ core-ng/src/main/java/core/framework/impl/db/RepositoryImpl.java
@@ -1,5 +1,6 @@
 package core.framework.impl.db;
 
+import core.framework.api.db.Query;
 import core.framework.api.db.Repository;
 import core.framework.api.log.ActionLogContext;
 import core.framework.api.log.Markers;
@@ -38,26 +39,10 @@ public final class RepositoryImpl<T> implements Repository<T> {
     }
 
     @Override
-    public List<T> selectAll() {
+    public List<T> select(Query query) {
         StopWatch watch = new StopWatch();
-        String sql = selectQuery.selectAll;
-        List<T> results = null;
-        try {
-            results = database.operation.select(sql, rowMapper, null);
-            return results;
-        } finally {
-            long elapsedTime = watch.elapsedTime();
-            ActionLogContext.track("db", elapsedTime);
-            logger.debug("selectAll, sql={}, elapsedTime={}", sql, elapsedTime);
-            checkSlowQuery(elapsedTime);
-            if (results != null) checkTooManyRows(results.size());
-        }
-    }
-
-    @Override
-    public List<T> select(String whereClause, Object... params) {
-        StopWatch watch = new StopWatch();
-        String sql = selectQuery.where(whereClause);
+        String sql = selectQuery.sql(query.where, query.skip, query.limit);
+        Object[] params = selectQuery.params(query);
         List<T> results = null;
         try {
             results = database.operation.select(sql, rowMapper, params);
@@ -72,9 +57,9 @@ public final class RepositoryImpl<T> implements Repository<T> {
     }
 
     @Override
-    public Optional<T> selectOne(String whereClause, Object... params) {
+    public Optional<T> selectOne(String where, Object... params) {
         StopWatch watch = new StopWatch();
-        String sql = selectQuery.where(whereClause);
+        String sql = selectQuery.sql(where, null, null);
         try {
             return database.operation.selectOne(sql, rowMapper, params);
         } finally {
@@ -123,7 +108,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
         try {
             int updatedRows = database.operation.update(query.sql, query.params);
             if (updatedRows != 1)
-                logger.warn(Markers.errorType("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
+                logger.warn(Markers.errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("db", elapsedTime);
@@ -138,7 +123,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
         try {
             int deletedRows = database.operation.update(deleteSQL, primaryKeys);
             if (deletedRows != 1)
-                logger.warn(Markers.errorType("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", deletedRows);
+                logger.warn(Markers.errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", deletedRows);
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("db", elapsedTime);
@@ -180,7 +165,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
             int[] deletedRows = database.operation.batchUpdate(deleteSQL, params);
             for (int deletedRow : deletedRows) {
                 if (deletedRow != 1) {
-                    logger.warn(Markers.errorType("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", Arrays.toString(deletedRows));
+                    logger.warn(Markers.errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", Arrays.toString(deletedRows));
                     break;
                 }
             }
@@ -194,13 +179,13 @@ public final class RepositoryImpl<T> implements Repository<T> {
 
     private void checkTooManyRows(int size) {
         if (size > database.tooManyRowsReturnedThreshold) {
-            logger.warn(Markers.errorType("TOO_MANY_ROWS_RETURNED"), "too many rows returned, returnedRows={}", size);
+            logger.warn(Markers.errorCode("TOO_MANY_ROWS_RETURNED"), "too many rows returned, returnedRows={}", size);
         }
     }
 
     private void checkSlowQuery(long elapsedTime) {
         if (elapsedTime > database.slowQueryThresholdInMs) {
-            logger.warn(Markers.errorType("SLOW_QUERY"), "slow db query, elapsedTime={}", elapsedTime);
+            logger.warn(Markers.errorCode("SLOW_QUERY"), "slow db query, elapsedTime={}", elapsedTime);
         }
     }
 }