+++ core-ng/src/main/java/core/framework/impl/log/ActionLog.java
@@ -1,11 +1,15 @@
 package core.framework.impl.log;
 
+import core.framework.api.log.Markers;
 import core.framework.api.util.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.slf4j.Marker;
 
 import java.time.Duration;
 import java.time.Instant;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -14,6 +18,7 @@ import java.util.UUID;
  * @author neo
  */
 public final class ActionLog {
+    private static final int MAX_TRACE_HOLD_SIZE = 5000;
     public final String id = UUID.randomUUID().toString();
     final Instant startTime = Instant.now();
     final Map<String, String> context = Maps.newLinkedHashMap();
@@ -24,23 +29,35 @@ public final class ActionLog {
     String refId;
     String errorMessage;
     long elapsed;
-    LogLevel result = LogLevel.INFO;
-    private String errorType;
-
-    void end() {
-        elapsed = Duration.between(startTime, Instant.now()).toMillis();
-    }
+    List<LogEvent> events = new LinkedList<>();
+    private LogLevel result = LogLevel.INFO;
+    private String errorCode;
 
     void process(LogEvent event) {
-        if (event.level.value > result.value) result = event.level;
-        String errorType = event.errorType();
-        if (errorType != null) {
-            this.errorType = errorType;
-            this.errorMessage = event.message();
+        if (event.level.value > result.value) {
+            result = event.level;
+            errorCode = event.errorCode(); // only update error type/message if level raised, so error type will be first WARN or first ERROR
+            errorMessage = event.message();
         }
-        if (event.trace()) {
-            trace = true;
+
+        if (events.size() < MAX_TRACE_HOLD_SIZE) {
+            events.add(event);
         }
+    }
+
+    void end() {
+        if (events.size() == MAX_TRACE_HOLD_SIZE) {
+            Marker marker = Markers.errorCode("TRACE_LOG_TOO_LONG");
+            String message = "reached max holding size of trace log, please contact arch team";
+            if (result.value < LogLevel.WARN.value) {       // not hide existing warn/error if there is already one
+                result = LogLevel.WARN;
+                errorCode = marker.getName();
+                errorMessage = message;
+            }
+            LogEvent warning = new LogEvent(logger.getName(), marker, LogLevel.WARN, message, null, null);
+            events.add(warning);
+        }
+        elapsed = Duration.between(startTime, Instant.now()).toMillis();
     }
 
     String result() {
@@ -50,8 +67,12 @@ public final class ActionLog {
         return String.valueOf(result);
     }
 
-    String errorType() {
-        if (errorType != null) return errorType;
+    boolean flushTraceLog() {
+        return trace || result.value >= LogLevel.WARN.value;
+    }
+
+    String errorCode() {
+        if (errorCode != null) return errorCode;
         if (result.value >= LogLevel.WARN.value) return "UNASSIGNED";
         return null;
     }