+++ core-ng/src/main/java/core/framework/impl/log/LogManager.java
@@ -1,20 +1,22 @@
 package core.framework.impl.log;
 
+import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Markers;
 import core.framework.api.log.Warning;
-import core.framework.impl.log.marker.ErrorTypeMarker;
 import org.slf4j.Logger;
+import org.slf4j.Marker;
 
 /**
  * @author neo
  */
 public final class LogManager {
     public final String appName;
+
     private final ThreadLocal<ActionLog> actionLog = new ThreadLocal<>();
-    private final ThreadLocal<TraceLogger> traceLogger = new ThreadLocal<>();
-    public TraceLoggerFactory traceLoggerFactory;
-    public LogForwarder logForwarder;
+
     public ActionLogger actionLogger;
+    public TraceLogger traceLogger;
+    public LogForwarder logForwarder;
     Logger logger;
 
     public LogManager() {
@@ -25,11 +27,8 @@ public final class LogManager {
         ActionLog actionLog = new ActionLog();
         this.actionLog.set(actionLog);
 
-        if (traceLoggerFactory != null) {
-            traceLogger.set(traceLoggerFactory.create(actionLog, logForwarder));
-            logger.debug(message);     // if trace log is disabled, then no need to process debug log
-            logger.debug("[context] id={}", actionLog.id);
-        }
+        logger.debug(message);
+        logger.debug("[context] id={}", actionLog.id);
     }
 
     public void end(String message) {
@@ -41,23 +40,13 @@ public final class LogManager {
         actionLog.end();
 
         if (actionLogger != null) actionLogger.write(actionLog);
-        if (logForwarder != null) logForwarder.forwardActionLog(actionLog);
-
-        if (traceLoggerFactory != null) {
-            TraceLogger traceLogger = this.traceLogger.get();
-            this.traceLogger.remove();
-            traceLogger.close();
-        }
+        if (traceLogger != null) traceLogger.write(actionLog);
+        if (logForwarder != null) logForwarder.forward(actionLog);
     }
 
     public void process(LogEvent event) {
-        if (event.isWarningOrError() || event.trace()) {
-            ActionLog actionLog = currentActionLog();
-            if (actionLog != null) actionLog.process(event);    // process is called by loggerImpl.log, begin() may not be called
-        }
-
-        TraceLogger traceLogger = this.traceLogger.get();
-        if (traceLogger != null) traceLogger.process(event);
+        ActionLog actionLog = currentActionLog();
+        if (actionLog != null) actionLog.process(event);    // process is called by loggerImpl.log, begin() may not be called before
     }
 
     public void start() {
@@ -75,17 +64,17 @@ public final class LogManager {
 
     public void triggerTraceLog() {
         ActionLog actionLog = currentActionLog();   // actionLog should not be null, logManager.begin should always be called before triggerTraceLog
-        logger.debug(Markers.TRACE, "trigger trace log, id={}, action={}", actionLog.id, actionLog.action);
+        actionLog.trace = true;
     }
 
     public void logError(Throwable e) {
-        // write exception first, to avoid hiding it by other mistake
         String errorMessage = e.getMessage();
-        ErrorTypeMarker errorType = Markers.errorType(e.getClass().getCanonicalName());
+        String errorCode = e instanceof ErrorCode ? ((ErrorCode) e).errorCode() : e.getClass().getCanonicalName();
+        Marker marker = Markers.errorCode(errorCode);
         if (e.getClass().isAnnotationPresent(Warning.class)) {
-            logger.warn(errorType, errorMessage, e);
+            logger.warn(marker, errorMessage, e);
         } else {
-            logger.error(errorType, errorMessage, e);
+            logger.error(marker, errorMessage, e);
         }
     }
 }