+++ log-processor/src/main/java/core/log/LogProcessorApp.java
@@ -2,12 +2,10 @@ package core.log;
 
 import core.framework.api.App;
 import core.framework.api.module.SystemModule;
-import core.framework.impl.log.queue.ActionLogMessage;
-import core.framework.impl.log.queue.TraceLogMessage;
+import core.framework.impl.log.queue.ActionLogMessages;
 import core.log.domain.ActionLogDocument;
 import core.log.domain.TraceLogDocument;
-import core.log.queue.ActionLogMessageHandler;
-import core.log.queue.TraceLogMessageHandler;
+import core.log.queue.ActionLogMessagesHandler;
 
 /**
  * @author neo
@@ -20,13 +18,10 @@ public class LogProcessorApp extends App {
         search().type(ActionLogDocument.class);
         search().type(TraceLogDocument.class);
 
-        // with typical t2.medium/t2.large setup, 15/10 are optimal number to keep CPU around 50%~70% with large amount of messages
+        // with typical t2.medium/t2.large setup, by bulkIndex, all requests will be queued up in ES in bulk queue,
+        // so no need to increase concurrency here
         queue().subscribe("rabbitmq://queue/action-log-queue")
-            .handle(ActionLogMessage.class, bind(ActionLogMessageHandler.class))
-            .maxConcurrentHandlers(15);
-
-        queue().subscribe("rabbitmq://queue/trace-log-queue")
-            .handle(TraceLogMessage.class, bind(TraceLogMessageHandler.class))
-            .maxConcurrentHandlers(1);  // trace message is to append to existing message id, must be handled in single thread, otherwise ES will lost trace log
+            .handle(ActionLogMessages.class, bind(ActionLogMessagesHandler.class))
+            .maxConcurrentHandlers(2);
     }
 }