+++ log-processor/src/main/java/core/log/domain/ActionLogDocument.java
@@ -29,10 +29,10 @@ public class ActionLogDocument {
     public String refId;
     @XmlElement(name = "action")
     public String action;
+    @XmlElement(name = "error_code")
+    public String errorCode;
     @XmlElement(name = "error_message")
     public String errorMessage;
-    @XmlElement(name = "exception_class")
-    public String errorType;
     @XmlElement(name = "elapsed")
     public Long elapsed;
     @XmlElement(name = "context")